Fixed bug in HugeArray Span and Fill methods. Added HugeArray benchmark.

+++ jemalloc.Benchmarks/JemBenchmark.cs
@@ -53,10 +53,10 @@ namespace jemalloc.Benchmarks
                     return JemUtil.ValToGenericStruct<Int32, TData>(Int32.MaxValue / 2);
 
                 case UInt64 v:
-                    return JemUtil.ValToGenericStruct<UInt64, TData>(UInt64.MaxValue / 2);
+                    return JemUtil.ValToGenericStruct<UInt64, TData>(Int32.MaxValue / 2);
 
                 case Int64 v:
-                    return JemUtil.ValToGenericStruct<Int64, TData>(Int64.MaxValue / 2);
+                    return JemUtil.ValToGenericStruct<Int64, TData>(Int32.MaxValue / 2);
 
                 default:
                     return value;
